+++ src/index.html
@@ -15,5 +15,6 @@
     <load src="./partials/main-section.html" />
 
     <script src="./main.js" type="module"></script>
+    <script src="./js/main-section/main-section.js" type="module"></script>
   </body>
 </html>
