+++ src/js/main-section/constants.js
@@ -1,4 +1,5 @@
 export const INVISIABLE_CLASS = 'invisiable';
+export const INDEX_PATH = '/goit-advancedjs-final-project/';
 
 export const FILTERS = {
   Muscles: 'muscles',
