+++ src/main.js
@@ -2,8 +2,6 @@ import './css/styles.css';
 import './js/menu.js';
 import './js/header-sticky.js';
 
-import './js/main-section/main-section.js';
-
 // NOTE: just sample how to use api methods/show errors
 import {
   getExercises,
